+++ src/unitMotions.ts
@@ -1,11 +1,12 @@
 import * as vscode from 'vscode';
 import { updateView } from './selectionMemory';
-import { IHash } from './util';
+import { clampedLineTranslate, IHash } from './util';
 
 interface MoveByArgs{
     unit?: string,
     select?: boolean,
     selectWhole?: boolean,
+    selectOneUnit?: boolean,
     value?: number
     boundary?: string,
 }
@@ -91,77 +92,74 @@ export function registerUnitMotions(context: vscode.ExtensionContext){
 }
 
 
+interface Range {
+    start?: vscode.Position,
+    end?: vscode.Position
+}
+
+function* singleLineUnitsForDoc(doc: vscode.TextDocument, start: vscode.Position, unit: RegExp, wrapAround: boolean, forwards: boolean){
+    let offset: number | undefined = start.character
+    for(const [line, i] of linesOf(doc, start, wrapAround || false, forwards)){
+        let matchesItr = regexMatches(unit, line, forwards, offset)
+        let matches = forwards ? matchesItr : Array.from(matchesItr).reverse()
+
+        yield* mapIter(matches, ([start, len]) => {return {
+            start: new vscode.Position(i, start),
+            end: new vscode.Position(i, start+len)
+        }})
+        offset = undefined
+    }
+}
+
+function* linesOf(doc: vscode.TextDocument, pos: vscode.Position,
+    wrap: boolean, forward: boolean): Generator<[string, number]>{
+
+    yield [doc.lineAt(pos).text, pos.line]
+    let line = pos.line + (forward ? 1 : -1)
+    while(forward ? line < doc.lineCount : line >= 0){
+        yield [doc.lineAt(line).text, line]
+        line += (forward ? 1 : -1)
+    }
+    if(wrap){
+        line = forward ? 0 : doc.lineCount - 1
+        while(forward ? line < doc.lineCount : line > 0){
+            yield [doc.lineAt(line).text, line]
+            line += (forward ? 1 : -1)
+        }
+    }
+}
+
+function* regexMatches(matcher: RegExp, line: string, forward: boolean, offset: number | undefined): Generator<[number, number]>{
+    matcher.lastIndex = 0
+    let match = matcher.exec(line)
+    while(match){
+        if(offset !== undefined && !forward && match.index > offset){ return }
+        if(offset === undefined || !forward || match.index+match[0].length > offset)
+            yield [match.index, match[0].length]
+        let newmatch = matcher.exec(line)
+        if(newmatch && newmatch.index > match.index){
+            match = newmatch
+        }else{
+            match = null
+        }
+    }
+}
+
+function* mapIter<T, R>(iter: Iterable<T>, fn: (x: T) => R){
+    for(const x of iter){
+        yield fn(x)
+    }
+}
+
+type Unit = RegExp | MultiLineUnit | string[]
 function unitsForDoc(document: vscode.TextDocument, from: vscode.Position,
-    boundary: Boundary, unit: RegExp | MultiLineUnit | string[], forward: boolean){
+    unit: Unit, forward: boolean){
 
     if(unit instanceof RegExp){
-        return singleLineUnitsForDoc(document, from, boundary, unit, forward);
+        return singleLineUnitsForDoc(document, from, unit, false, forward);
     }else{
-        return multiLineUnitsForDoc(document, from, boundary, unit as MultiLineUnit,
+        return multiLineUnitsForDoc(document, from, unit as MultiLineUnit,
             forward);
-    }
-}
-
-function* singleLineUnitsForDoc(document: vscode.TextDocument, from: vscode.Position,
-    boundary: Boundary, unit: RegExp, forward: boolean):
-    Generator<[vscode.Position, Boundary]>{
-
-    let line = from.line;
-    let char = from.character;
-    let str = document.lineAt(line).text;
-    if(forward){
-        for(let [pos, bound] of unitBoundaries(str,boundary,unit)){
-            if(pos < char) { continue; }
-            else { yield [new vscode.Position(line,pos), bound]; }
-        }
-        while(line < document.lineCount-1){
-            line++;
-            str = document.lineAt(line).text;
-            for(let [pos, bound] of unitBoundaries(str,boundary,unit)){
-                yield [new vscode.Position(line,pos), bound];
-            }
-        }
-        let finalChar = document.lineAt(document.lineCount-1).range.end.character;
-        yield [new vscode.Position(document.lineCount-1,finalChar),
-                boundary === Boundary.Both ? Boundary.End : boundary];
-    }else{
-        let positions = Array.from(unitBoundaries(str,boundary,unit));
-        if(positions.length > 0){
-            for(let [pos, bound] of positions.reverse()){
-                if(pos > char) { continue; }
-                else { yield [new vscode.Position(line,pos), bound]; }
-            }
-        }
-        while(line > 0){
-            line--;
-            str = document.lineAt(line).text;
-            let positions = Array.from(unitBoundaries(str,boundary,unit));
-            for(let [pos, bound] of positions.reverse()){
-                yield [new vscode.Position(line,pos), bound];
-            }
-        }
-        yield [new vscode.Position(0,0),
-            boundary === Boundary.Both ? Boundary.Start : boundary];
-    }
-}
-
-
-function* unitBoundaries(text: string,boundary: Boundary, unit: RegExp): Generator<[number, Boundary]>{
-    let reg = RegExp(unit);
-    reg.lastIndex = 0;
-    let match = reg.exec(text);
-    let boundaries: number[] = [];
-
-    while(match){
-        if(boundary === Boundary.Start){
-            yield [match.index, Boundary.Start];
-        }else if(boundary === Boundary.End){
-            yield [match.index + match[0].length, Boundary.End];
-        }else if(boundary === Boundary.Both){
-            yield [match.index, Boundary.Start];
-            yield [match.index + match[0].length, Boundary.End];
-        }
-        match = reg.exec(text);
     }
 }
 
@@ -177,71 +175,106 @@ export function first<T>(x: Iterable<T>): T | undefined {
 
 enum Boundary { Start, End, Both }
 
-interface MultiLineUnit { regexs: RegExp[], }
+interface MultiLineUnit { regexs: RegExp[] }
+
+function* docLines(document: vscode.TextDocument, from: vscode.Position, forward: boolean): Generator<[number, string]>{
+    yield [from.line, document.lineAt(from).text];
+    if(forward){
+        while(from.line+1 < document.lineCount){
+            from = new vscode.Position(from.line+1, 0);
+            yield [from.line, document.lineAt(from).text];
+        }
+    }else{
+        while(from.line > 0){
+            from = new vscode.Position(from.line-1, 0);
+            yield [from.line, document.lineAt(from).text];
+        }
+    }
+}
+
+function* singleRegexUnitsForDoc(document: vscode.TextDocument, from: vscode.Position,
+    unit: MultiLineUnit, forward: boolean){
+    
+    let first_match: undefined | null | number = undefined;
+    function closeMatch(line: number){
+        if(first_match !== undefined){
+            let startLine: null | number = forward ? first_match : line+1;
+            let endLine: null | number = forward ? line-1 : first_match;
+            let endCol = endLine === null ? null : document.lineAt(new vscode.Position(endLine, 0)).range.end.character;
+            return {
+                start: startLine !== null ? new vscode.Position(startLine, 0) : undefined,
+                end: endLine !== null && endCol !== null ? new vscode.Position(endLine, endCol) : undefined,
+            };
+        }
+    }
+    let first = true
+    for(let [line, text] of docLines(document, from, forward)){
+        if(unit.regexs[0].test(text)){
+            if(first){
+                // if the very first line matches, the `from` position could
+                // fall in the middle of a sequence of matching lines we
+                // represent the fact that we dont' know where this sequence
+                // starts using `null`
+                first_match = null;
+            }else if(first_match === undefined){
+                first_match = line;
+            }
+        }else{
+            let result = closeMatch(line)
+            if(result !== undefined){
+                first_match = undefined
+                yield result
+            } 
+        }
+        first = false
+    }
+    let result = closeMatch(forward ? document.lineCount : -1);
+    if(result !== undefined) yield result;
+}
+
+function* multiRegexUnitsForDoc(document: vscode.TextDocument, from: vscode.Position,
+    unit: MultiLineUnit, forward: boolean){
+    
+    let buffer: [number, string][] = [];
+    let doesMatch
+    if(forward){
+        doesMatch = (unit: MultiLineUnit, buffer: [number, string][]) => {
+            return unit.regexs.every((x, i) => x.test(buffer[i][1]))
+        }
+    }else{
+        doesMatch = (unit: MultiLineUnit, buffer: [number, string][]) => {
+            return unit.regexs.every((x, i) => x.test(buffer[buffer.length - (i+1)][1]));
+        }
+    }
+
+    let start_from = clampedLineTranslate(from, document, -unit.regexs.length+1);
+    for(let [line, text] of docLines(document, start_from, forward)){
+        buffer.push([line, text])
+        if(buffer.length > unit.regexs.length){
+            buffer.shift();
+        }
+        if(buffer.length === unit.regexs.length){
+            if(doesMatch(unit, buffer)){
+                let startLine = forward ? buffer[0][0] : buffer[buffer.length-1][0]
+                let endLine =  !forward ? buffer[0][0] : buffer[buffer.length-1][0]
+                let endCol = document.lineAt(endLine).range.end.character
+                yield {
+                    start: new vscode.Position(startLine, 0),
+                    end: new vscode.Position(endLine, endCol)
+                }
+            }
+        }
+    }
+}
 
 function* multiLineUnitsForDoc(document: vscode.TextDocument, from: vscode.Position,
-    boundary: Boundary, unit: MultiLineUnit, forward: boolean):
-    Generator<[vscode.Position, Boundary]>{
+    unit: MultiLineUnit, forward: boolean): Generator<Range>{
 
-    let lineNum = from.line + (unit.regexs.length-1)*(forward ? -1 : 1);
-    lineNum = Math.max(Math.min(document.lineCount, lineNum), 0);
-    let start = lineNum;
-    let curLinesToMatch: string[] = [];
-    let lines: string[] = [];
-    let lastTest: boolean | undefined = undefined;
-    let finalBoundary = forward ? Boundary.End : Boundary.Start;
-    if(lineNum >= document.lineCount && !forward){
-        lineNum = document.lineCount-1;
+    if(unit.regexs.length > 1){
+        yield* multiRegexUnitsForDoc(document, from, unit, forward);
+    }else{
+        yield* singleRegexUnitsForDoc(document, from, unit, forward);
     }
-    while(forward ? lineNum < document.lineCount : lineNum >= 0){
-        let line = document.lineAt(lineNum).text;
-        curLinesToMatch.push(line);
-        if(curLinesToMatch.length > unit.regexs.length){
-            curLinesToMatch.shift();
-        }
-        let ismatch = curLinesToMatch.length === unit.regexs.length ?
-            forward ? unit.regexs.every((x,i) => x.test(curLinesToMatch[i])) :
-            unit.regexs.
-                every((x,i) => x.test(curLinesToMatch[curLinesToMatch.length-(i+1)])) :
-            false;
-        if(ismatch){ lines.push(line); }
-        if(lastTest !== undefined && ismatch !== lastTest){
-            lastTest = ismatch;
-            let pos;
-            if(ismatch === forward && boundary !== Boundary.End){
-                pos = new vscode.Position(forward ?
-                    (boundary === Boundary.Both ? lineNum : lineNum-1) :
-                    lineNum + 1,0);
-                finalBoundary = Boundary.End;
-                if(forward ? pos.line >= from.line : pos.line <= from.line){
-                    yield [pos, Boundary.Start];
-                }
-            }
-            if(ismatch !== forward && boundary !== Boundary.Start){
-                let line = forward ?
-                    (boundary === Boundary.Both ? lineNum-unit.regexs.length : lineNum-unit.regexs.length+1) :
-                    lineNum+unit.regexs.length-1;
-                let endchar = document.lineAt(line).range.end.character;
-                pos = new vscode.Position(line,endchar);
-                finalBoundary = Boundary.Start;
-                if(forward ? pos.line >= from.line : pos.line <= from.line){
-                    yield [pos, Boundary.End];
-                }
-            }
-            lines = [];
-            start = forward ? lineNum+1 : lineNum-1;
-        }
-        lastTest = ismatch;
-        forward ? lineNum++ : lineNum--;
-    }
-    // handle boundaries at start and end of document
-    let documentBoundary = forward ?
-        new vscode.Position(document.lineCount-1,
-            document.lineAt(document.lineCount-1).range.end.character) :
-        new vscode.Position(0,0);
-    yield [documentBoundary, boundary !== Boundary.Both ? boundary :
-        finalBoundary];
-    return;
 }
 
 function unitNameToRegex(editor: vscode.TextEditor, name?: string){
@@ -253,6 +286,250 @@ function unitNameToRegex(editor: vscode.TextEditor, name?: string){
         allUnits[id] === undefined ? allUnits['[GENERIC]'][name] :
         allUnits[id][name] || allUnits['[GENERIC]'][name];
 }
+
+function toBoundary(args: {boundary?: string}) {
+    if(args.boundary === undefined){
+        return Boundary.Start;
+    }else if(args.boundary === 'start'){
+        return Boundary.Start;
+    }else if(args.boundary === 'end'){
+        return Boundary.End;
+    }else if(args.boundary === 'both'){
+        return Boundary.Both;
+    }else{
+        vscode.window.showErrorMessage("Unexpected value for boundary argument: '"+
+            args.boundary+"'.");
+        return undefined
+    }
+}
+
+// this handles unit cleanup when looking for two boundaries, it handles two
+// problesm:
+// 1. when you look for units in one direction sometime you miss the start (or
+//    end) of a unit you're in the middle of (for multi-line units in particular).
+//    If we want to resolve the boundaries of a unit, we need to look 
+//    backwards from the starting position.
+// 2. when you get to the end of the document you need to treat the edge of the
+//    document as unit boundaries to make things work out cleanly for
+//    'start-only' and `end-only` boundary resolution
+function* resolveUnitBoundaries(resolve: Boundary | undefined,
+    units: Generator<Range>, document: vscode.TextDocument,
+    from: vscode.Position, unit: Unit,
+    forward: boolean): Generator<Range> {
+
+    function* resolveUnit(first_unit: Range | undefined){
+        let backwards = unitsForDoc(document, from, unit, !forward)
+        let foundUnit = false
+        function* resolveHelper(back: Range){
+            if(resolve === Boundary.Start){
+                if(back.start){
+                    yield back;
+                    if(first_unit?.start){ yield first_unit }
+                    else{
+                        let next = first(units)
+                        if(next?.start){ yield next; }
+                        else yield {start: forward ? lastPosition(document) : firstPosition()};
+                    }
+                    foundUnit = true
+                }
+            }else if(resolve === Boundary.End){
+                if(back.end){
+                    yield back;
+                    if(first_unit?.end) yield first_unit
+                    else{
+                        let next = first(units);
+                        if(next?.end) yield next;
+                        else yield {end: forward ? lastPosition(document) : firstPosition()};
+                    }
+                    foundUnit = true
+                }
+            }else if(back.start){
+                if(back.end){ 
+                    foundUnit = true;  
+                    yield back; 
+                } else if(first_unit?.end){
+                    foundUnit = true;
+                    yield {start: back.start, end: first_unit.end}
+                }
+                if(first_unit?.end && first_unit?.start) {
+                    yield first_unit
+                    foundUnit = true
+                }
+            }
+        }
+        for(let back of backwards){
+            yield* resolveHelper(back);
+            if(foundUnit) return;
+        }
+
+        if(resolve !== Boundary.Both){
+            if(forward){
+                let first = firstPosition()
+                yield* resolveHelper({start: first, end: first});
+            }else{
+                let last = lastPosition(document)
+                yield* resolveHelper({start: last, end: last});
+            }
+        }
+    }
+
+    let first_unit = first(units)
+    if (first_unit) {
+        if(forward){
+            if(resolve === Boundary.End ? 
+                !first_unit.end || first_unit.end.isAfter(from) : 
+                !first_unit.start || first_unit.start.isAfter(from)) {
+                yield* resolveUnit(first_unit);
+            }else yield first_unit
+        }else if(resolve === Boundary.Start ? 
+                 !first_unit.start || first_unit.start.isBefore(from) :
+                 !first_unit.end   || first_unit.end.isBefore(from)){
+            yield* resolveUnit(first_unit);
+        }else{
+            yield first_unit;
+        }
+    }else {
+        yield* resolveUnit(first_unit);
+    }
+    yield* units;
+    if(resolve !== Boundary.Both){
+        if(forward){
+            let last = lastPosition(document)
+            yield {start: last, end: last};
+        }else{
+            let first = firstPosition()
+            yield {start: first, end: first};
+        }
+    }
+}
+
+function lastPosition(document: vscode.TextDocument){
+    let last = document.lineCount-1;
+    let endCol = document.lineAt(last).range.end.character;
+    return new vscode.Position(last, endCol);
+}
+function firstPosition(){
+    return new vscode.Position(0, 0);
+}
+
+function moveBy(editor: vscode.TextEditor,args: MoveByArgs){
+    let unit = unitNameToRegex(editor, args.unit);
+    let forward = args.value === undefined ? true : args.value > 0;
+    let holdSelect = args.select === undefined ? false : args.select;
+    let selectWholeUnit = args.selectWhole === undefined ? false : args.selectWhole;
+    let selectOneUnit = args.selectOneUnit;
+
+    let boundary = toBoundary(args);
+    if(boundary === undefined){ return (sel: vscode.Selection) => sel; }
+    let steps = args.value === undefined ? 1 : Math.abs(args.value);
+    if(steps === 0) { return (select: vscode.Selection) => select; }
+
+    // translate a sequence of units (regex start and stop boundaries)
+    // to a sequence of selection points (where we extend the active selection
+    // point to a new unit boundary at each step)
+    function* selectBoundaries(xs: Generator<Range>, start: vscode.Position | undefined){
+        function withStart(x: vscode.Position){
+            if(start) return new vscode.Selection(start, x);
+            else return new vscode.Selection(x, x);
+        }
+        for(let x of xs){
+            if(boundary === Boundary.Start){
+                if(x.start) yield withStart(x.start)
+            }else if(boundary === Boundary.End){
+                if(x.end) yield withStart(x.end);
+            }else{
+                if(x.start) yield withStart(x.start);
+                if(x.end) yield withStart(x.end);
+            }
+        }
+        // treat the edge of the document as a boundary as well
+        if(forward){
+            yield withStart(lastPosition(editor.document));
+        }else{
+            yield withStart(firstPosition());
+        }
+    }
+
+    // translate a sequence of units (regex start and stop boundaries)
+    // to a sequence of selections: the selections surround a single
+    // unit around from start-to-start, end-to-end or start-to-end
+    function* selectUnits(xs: Generator<Range>, forward: boolean){
+        let last: vscode.Position | undefined | null = null;
+        let current: vscode.Position | undefined | null;
+        for(let x of xs){
+            if(boundary !== Boundary.Both){
+                last = current;
+                current = boundary === Boundary.Start ? x.start : x.end
+                if(current && last){
+                    yield new vscode.Selection(last, current);
+                }
+            }else{
+                if(!x.start || !x.end) throw new Error("Unexpected missing range bound")
+                if(forward) yield new vscode.Selection(x.start, x.end); 
+                else yield new vscode.Selection(x.start, x.end);  
+            }
+        }
+        if(boundary !== Boundary.Both){
+            if(forward){
+                if(current === undefined) throw new Error("Unexepcted missing range bound")
+                else if(current !== null)
+                    yield new vscode.Selection(current, lastPosition(editor.document));
+            }else{
+                if(current === undefined) throw new Error("Unexepcted missing range bound")
+                else if(current !== null)
+                    yield new vscode.Selection(current, firstPosition());
+            }
+        }
+    }
+
+    // return a function that modifies each selection in turn
+    // (it will be applied to all selections)
+    return (select: vscode.Selection) => {
+        let units = unitsForDoc(editor.document, select.active, unit, forward);
+        let selections
+        if(selectWholeUnit){
+            // if we are selecting whole units we need to use `resolveUnitBoundaries`
+            // to look possibly backwards from the starting position to resolve
+            // the start of a unit
+            let resolved = resolveUnitBoundaries(boundary, units, editor.document, select.active, unit, forward);
+            selections = selectUnits(resolved, forward);
+        }else{
+            selections = selectBoundaries(units, holdSelect ? select.anchor : undefined);
+        }
+        let count = 0;
+        let lastsel;
+        let startSel: vscode.Position | undefined = undefined
+        for(let sel of selections){
+            if(count > 0 || !boundsMatch(sel, select)){
+                if(forward ? sel.end.isAfter(select.active) : sel.start.isBefore(select.active))
+                    count += 1;
+                if(count === 1){ startSel = sel.anchor; }
+            }
+            if (count >= steps){
+                if(!selectOneUnit && startSel && selectWholeUnit){
+                    return new vscode.Selection(startSel, sel.active);
+                }else{
+                    return sel;
+                }
+            }
+            lastsel = sel;
+        }
+        if(count > 0 && lastsel){ 
+            if(!selectOneUnit && startSel){
+                return new vscode.Selection(startSel, lastsel.active);
+            }else{
+                return lastsel; 
+            }
+        }
+        return select;
+    };
+};
+
+function boundsMatch(x: vscode.Selection, y: vscode.Selection){
+    return (x.start.isEqual(y.start) && x.end.isEqual(y.end)) ||
+        (x.end.isEqual(y.start) && x.start.isEqual(y.end))
+}
+
 function narrowTo(editor: vscode.TextEditor, args: NarrowByArgs): (select: vscode.Selection) => vscode.Selection {
     let unit = unitNameToRegex(editor, args.unit);
     let thenNarrow = args.then === undefined ? undefined :
@@ -262,112 +539,40 @@ function narrowTo(editor: vscode.TextEditor, args: NarrowByArgs): (select: vscod
                 args.thenBoundary
         });
 
-    let boundary: Boundary;
-    if(args.boundary === undefined){
-        boundary = Boundary.Both;
-    }else if(args.boundary === 'start'){
-        boundary = Boundary.Start;
-    }else if(args.boundary === 'end'){
-        boundary = Boundary.End;
-    }else if(args.boundary === 'both'){
-        boundary = Boundary.Both;
-    }else{
-        vscode.window.showErrorMessage("Unexpected value for boundary argument: '"+args.boundary+"'.");
-        return (select: vscode.Selection) => select;
-    }
+    let boundary = toBoundary(args);
+    if(boundary === undefined){ return (sel: vscode.Selection) => sel; }
 
     return (select: vscode.Selection) => {
         if(select.anchor.isEqual(select.active)){
             return select;
         }
-        let starts = unitsForDoc(editor.document,select.start,
-            boundary === Boundary.Both ? Boundary.Start : boundary,
-            unit,true);
-        let step = first(starts);
-        let start = step === undefined ? select.start : step[0];
-
-        let stops = unitsForDoc(editor.document,select.end,
-            boundary === Boundary.Both ? Boundary.End : boundary,
-            unit,false);
-        step = first(stops);
-        let stop = step === undefined ? select.end : step[0];
-
-        if(stop.isEqual(select.end) && start.isEqual(select.start)){
-            if(thenNarrow){ return thenNarrow(select); }
+        let starts = unitsForDoc(editor.document,select.start,unit,true);
+        let start: vscode.Position | undefined;
+        for(let s of starts){
+            if(s.start?.isAfterOrEqual(select.end)) break;
+            if(s.start?.isAfterOrEqual(select.start)){
+                start = s.start
+                break;
+            }
         }
-        if(stop.isBefore(select.start) || start.isAfter(select.end)){
-            if(thenNarrow){ return thenNarrow(select); }
-            else{ return select; }
+        let stops = unitsForDoc(editor.document,select.end,unit,false);
+        let stop: vscode.Position | undefined;
+        for(let s of stops){
+            if(s.end?.isBeforeOrEqual(select.start)) break;
+            if(s.end?.isBeforeOrEqual(select.end)){
+                stop = s.end;
+                break;
+            } 
         }
+        if(!stop || !start){
+            if(thenNarrow){ return thenNarrow(select);}
+            return select;
+        }
+        let result;
         if(select.anchor.isBefore(select.active)){
             return new vscode.Selection(start,stop);
         }else{
             return new vscode.Selection(stop,start);
-        }
-    };
-}
-
-function moveBy(editor: vscode.TextEditor,args: MoveByArgs){
-    let unit = unitNameToRegex(editor, args.unit);
-    let forward = args.value === undefined ? true : args.value > 0;
-    let holdSelect = args.select === undefined ? false : args.select;
-    let selectWholeUnit = args.selectWhole === undefined ? false : args.selectWhole;
-
-    let boundary: Boundary;
-    if(args.boundary === undefined){
-        boundary = Boundary.Start;
-    }else if(args.boundary === 'start'){
-        boundary = Boundary.Start;
-    }else if(args.boundary === 'end'){
-        boundary = Boundary.End;
-    }else if(args.boundary === 'both'){
-        boundary = Boundary.Both;
-    }else{
-        vscode.window.showErrorMessage("Unexpected value for boundary argument: '"+
-            args.boundary+"'.");
-        return (select: vscode.Selection) => select;
-    }
-    let steps = args.value === undefined ? 1 : Math.abs(args.value);
-    if(steps === 0) { return (select: vscode.Selection) => select; }
-
-    return (select: vscode.Selection) => {
-        // TODO: if its a default word, we take advantage
-        // of language specific word definitions
-        let start: vscode.Position | undefined = undefined;
-        if(selectWholeUnit){
-            let units = unitsForDoc(editor.document,select.active,boundary,
-                unit,!forward);
-            let value = first(units);
-            if(value !== undefined) { [start] = value; }
-        }else if(holdSelect){
-            start = select.anchor;
-        }
-
-        let units = unitsForDoc(editor.document,select.active,boundary,
-            unit,forward);
-        let count = 0;
-        let pos = select.active;
-        let bound: Boundary;
-        for([pos, bound] of units){
-            if(forward ? bound === Boundary.Start : bound === Boundary.End){
-                if(selectWholeUnit && boundary === Boundary.Both){
-                    start = pos;
-                }
-            }
-            if(!pos.isEqual(select.active)){
-                if(selectWholeUnit && boundary === Boundary.Both){
-                    if(forward ? bound === Boundary.End :
-                                              bound === Boundary.Start){
-                        count++;
-                    }
-                }else{ count++; }
-            }
-            if(count === steps) { break; }
-        }
-        if(start){
-            return new vscode.Selection(start,pos);
-        }else{
-            return new vscode.Selection(pos,pos);
         }
     };
 }